+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-08-14 03:13:19
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 7366dc403a884fbd8008139bbac372f6f3a599ad
+
+## [ForwardWidgets-2kuai] - 2025-08-14 03:13:19
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 979c1ad54d65f1f27f6018f224b9fde253290472
+
+## [ForwardWidgets-pack1r] - 2025-08-14 03:13:19
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 0cf5a9a386b2c1f9faece6e1c6dcd340f86b1d74
+
+
 ## [ForwardWidgets-huangxd] - 2025-08-13 03:12:48
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main